Co-locate WithWpCompatOverlaySlot decorator with playground stories
Moves the decorator out of `@wordpress/ui` (where it had ended up to
avoid a `packages/ui` → `storybook/` cross-boundary dependency for the
per-component Tooltip story) and into the same directory as the only
remaining consumer, the cross-library Playground story.

The dedicated `Inside WP Compat Overlay Slot` Tooltip story is removed
along with the move. The decorator mutates a window-level flag that's
process-wide across the Storybook preview iframe, so applying it to
per-component stories made the opt-in leak into every sibling story
rendered on the same autodocs page. Keeping the decorator next to the
single playground story that needs it confines the leak blast radius
to the playground; cross-library stacking remains the end-to-end
validator for the slot.
