+++ storybook/stories/playground/with-wp-compat-overlay-slot.tsx
@@ -1,0 +1,30 @@
+import { useLayoutEffect } from '@wordpress/element';
+import { useEnableWpCompatOverlaySlot } from '@wordpress/ui';
+import type { Decorator } from '@storybook/react-vite';
+
+/**
+ * Storybook decorator that opts the decorated story into the
+ * `@wordpress/ui` compat overlay slot. The cleanup clears the internal
+ * flag so navigating to an un-decorated story closes the gate again —
+ * `useEnableWpCompatOverlaySlot()` itself is intentionally one-way.
+ *
+ * Kept private to the playground stories: the flag is window-level, so
+ * applying the decorator to per-component stories would leak the opt-in
+ * into sibling stories on the same autodocs page.
+ */
+export const WithWpCompatOverlaySlot: Decorator = ( Story ) => {
+	useEnableWpCompatOverlaySlot();
+
+	useLayoutEffect( () => {
+		const internalWindow = window as {
+			__wpUiCompatOverlaySlotEnabled?: boolean;
+		};
+		internalWindow.__wpUiCompatOverlaySlotEnabled = true;
+		return () => {
+			delete ( window as { __wpUiCompatOverlaySlotEnabled?: boolean } )
+				.__wpUiCompatOverlaySlotEnabled;
+		};
+	}, [] );
+
+	return <Story />;
+};
